Map nb work: phy. units, paste range.
Replaced the hack of direct ticks to pixels to ticks to meters
and then meters to pixels for desired map size. Still needs work
but better than before.

Also worked on the source and destination rectangles for the map
image draw operation (sub-matrix copy). Whereas before I took a shortcut
and just selected a small enough range that it almost always fit,
now I actually compute the max possible range to copy. Of course, this
means, now I have to think harder about how to favor close pictures,
which I suppose is the next step on this one.