Mostly localization work. Some OpenCV.
Did some more experimentation with OpenCV, but then realized I don't
even have it built for Raspberry Pi on Python 3 yet and its not available
as binary, so had to go build it.

Most of the work in this is around trying to use wheel encoders alone
to track path on the track. It's going OK I guess but hitting some snag
and needed a checkpoint before continuing.